# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Pipeline (AI Orchestration) ─────────────────────────────────────────────
  /pipeline/run:
    post:
      operationId: runPipeline
      tags: [pipeline]
      summary: Run the MultiClaw-OpenClaw AI pipeline
      description: Triggers the AI agents pipeline to generate a full ads blueprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineInput"
      responses:
        "201":
          description: Pipeline started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRun"
...
  /pipeline/{pipelineId}:
    get:
      operationId: getPipelineRun
      tags: [pipeline]
      summary: Get a pipeline run and its results
      parameters:
        - name: pipelineId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunDetail"
        "404":
          description: Not found

  # ─── Blueprints ───────────────────────────────────────────────────────────────
...
  /adsets/{adsetId}/creatives:
    get:
      operationId: listCreatives
      tags: [creatives]
      summary: List creatives for an ad set
      parameters:
        - name: adsetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of creatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creative"
    post:
      operationId: createCreative
      tags: [creatives]
      summary: Create a creative for an ad set
      parameters:
        - name: adsetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreativeInput"
      responses:
        "201":
          description: Creative created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creative"
  /creatives/{creativeId}:
    get:
      operationId: getCreative
      tags: [creatives]
      summary: Get a creative
      parameters:
        - name: creativeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creative"
        "404":
          description: Not found
    patch:
      operationId: updateCreative
      tags: [creatives]
      summary: Update a creative
      parameters:
        - name: creativeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreativeUpdate"
      responses:
        "200":
          description: Creative updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creative"
    delete:
      operationId: deleteCreative
      tags: [creatives]
      summary: Delete a creative
      parameters:
        - name: creativeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Approvals ────────────────────────────────────────────────────────────────
...
  /memory/{businessId}:
    get:
      operationId: getBusinessMemory
      tags: [memory]
      summary: Get agent memory for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent memory for business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMemory"
        "404":
          description: Not found
components:
  schemas:
...
    Product:
      type: object
      required: [id, businessId, productName, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        productName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        promo:
          type: ["string", "null"]
        benefit:
          type: ["string", "null"]
        painPoint:
          type: ["string", "null"]
        proof:
          type: ["string", "null"]
        guarantee:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        landingPageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [productName]
      properties:
        productName:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
        promo:
          type: string
        benefit:
          type: string
        painPoint:
          type: string
        proof:
          type: string
        guarantee:
          type: string
        cta:
          type: string
        landingPageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        productName:
          type: string
        description:
          type: string
        price:
          type: number
        promo:
          type: string
        benefit:
          type: string
        painPoint:
          type: string
        proof:
          type: string
        guarantee:
          type: string
        cta:
          type: string
        landingPageUrl:
          type: string

    # ─── Pipeline ─────────────────────────────────────────────────────────────────
    PipelineInput:
      type: object
      required: [businessId, productId, campaignGoal, budget]
      properties:
        businessId:
          type: integer
        productId:
          type: integer
        campaignGoal:
          type: string
          description: "e.g. leads, sales, awareness, traffic"
        budget:
          type: number
          description: Daily budget in local currency
        targetLocation:
          type: string
        targetAgeMin:
          type: integer
        targetAgeMax:
          type: integer
        additionalContext:
          type: string
    PipelineRun:
      type: object
      required: [id, businessId, productId, status, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        productId:
          type: integer
        campaignGoal:
          type: string
        budget:
          type: number
        status:
          type: string
          description: "pending | running | completed | failed"
        agentLog:
          type: ["string", "null"]
        blueprintId:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    PipelineRunDetail:
      type: object
      required: [id, businessId, productId, status, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        productId:
          type: integer
        campaignGoal:
          type: string
        budget:
          type: number
        status:
          type: string
        agentLog:
          type: ["string", "null"]
        blueprintId:
          type: ["integer", "null"]
        blueprint:
          $ref: "#/components/schemas/Blueprint"
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]

    # ─── Blueprint ───────────────────────────────────────────────────────────────
    Blueprint:
      type: object
      required: [id, pipelineRunId, title, createdAt]
      properties:
        id:
          type: integer
        pipelineRunId:
          type: integer
        title:
          type: string
        businessContext:
          type: ["string", "null"]
          description: Business diagnosis section (JSON string)
        campaignStrategy:
          type: ["string", "null"]
          description: Campaign strategy recommendation (JSON string)
        audiencePlan:
          type: ["string", "null"]
          description: Audience intelligence blueprint (JSON string)
        offerStrategy:
          type: ["string", "null"]
          description: Offer engineering blueprint (JSON string)
        creativeBlueprint:
          type: ["string", "null"]
          description: Creative copy and angle blueprint (JSON string)
        budgetPlan:
          type: ["string", "null"]
          description: Budget and testing plan (JSON string)
        policyReview:
          type: ["string", "null"]
          description: Compliance and policy risk assessment (JSON string)
        conversionReadinessScore:
          type: ["integer", "null"]
          description: 0-100 score
        policyRiskScore:
          type: ["integer", "null"]
          description: 0-100 score
        creativeStrengthScore:
          type: ["integer", "null"]
          description: 0-100 score
        funnelFitScore:
          type: ["integer", "null"]
          description: 0-100 score
        approvalStatus:
          type: string
          description: "draft | needs_revision | approved | published"
        createdAt:
          type: string

    # ─── Campaign ────────────────────────────────────────────────────────────────
...
    Creative:
      type: object
      required: [id, adsetId, adName, createdAt]
      properties:
        id:
          type: integer
        adsetId:
          type: integer
        adName:
          type: string
        format:
          type: ["string", "null"]
          description: "single_image | video | carousel | collection"
        mediaUrl:
          type: ["string", "null"]
        primaryText:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        destinationUrl:
          type: ["string", "null"]
        utmParams:
          type: ["string", "null"]
        policyScore:
          type: ["integer", "null"]
          description: 0-100
        creativeScore:
          type: ["integer", "null"]
          description: 0-100
        angle:
          type: ["string", "null"]
          description: "problem_solution | benefit | testimonial | hard_selling | soft_selling | educational"
        createdAt:
          type: string
    CreativeInput:
      type: object
      required: [adName]
      properties:
        adName:
          type: string
          minLength: 1
        format:
          type: string
        mediaUrl:
          type: string
        primaryText:
          type: string
        headline:
          type: string
        description:
          type: string
        cta:
          type: string
        destinationUrl:
          type: string
        utmParams:
          type: string
        angle:
          type: string
    CreativeUpdate:
      type: object
      properties:
        adName:
          type: string
        format:
          type: string
        mediaUrl:
          type: string
        primaryText:
          type: string
        headline:
          type: string
        description:
          type: string
        cta:
          type: string
        destinationUrl:
          type: string
        utmParams:
          type: string
        angle:
          type: string

    # ─── Approvals ────────────────────────────────────────────────────────────────
...
    AgentMemory:
      type: object
      required: [id, businessId, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        winningAudience:
          type: ["string", "null"]
        winningCopy:
          type: ["string", "null"]
        winningHeadline:
          type: ["string", "null"]
        winningCta:
          type: ["string", "null"]
        winningOffer:
          type: ["string", "null"]
        failedPattern:
          type: ["string", "null"]
        policyIssue:
          type: ["string", "null"]
        optimizationHistory:
          type: ["string", "null"]
        learningSummary:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string