# lib/api-zod/src/generated/types/agentMemory.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentMemory {
  id: number;
  businessId: number;
  /** @nullable */
  winningAudience?: string | null;
  /** @nullable */
  winningCopy?: string | null;
  /** @nullable */
  winningHeadline?: string | null;
  /** @nullable */
  winningCta?: string | null;
  /** @nullable */
  winningOffer?: string | null;
  /** @nullable */
  failedPattern?: string | null;
  /** @nullable */
  policyIssue?: string | null;
  /** @nullable */
  optimizationHistory?: string | null;
  /** @nullable */
  learningSummary?: string | null;
  createdAt: string;
  updatedAt?: string;
}