# lib/api-zod/src/generated/types/creativeUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface CreativeUpdate {
  adName?: string;
  format?: string;
  mediaUrl?: string;
  primaryText?: string;
  headline?: string;
  description?: string;
  cta?: string;
  destinationUrl?: string;
  utmParams?: string;
  angle?: string;
}