# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  businessId: number;
  productName: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  promo?: string | null;
  /** @nullable */
  benefit?: string | null;
  /** @nullable */
  painPoint?: string | null;
  /** @nullable */
  proof?: string | null;
  /** @nullable */
  guarantee?: string | null;
  /** @nullable */
  cta?: string | null;
  /** @nullable */
  landingPageUrl?: string | null;
  createdAt: string;
}