# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  productName: string;
  description?: string;
  price?: number;
  promo?: string;
  benefit?: string;
  painPoint?: string;
  proof?: string;
  guarantee?: string;
  cta?: string;
  landingPageUrl?: string;
}